+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- `init` now accepts a list of module-level overrides
+- Replace `init` function with a builder pattern
 
 ## [0.2.0]