+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.0]
+
 ### Changed
 
 - Replace `init` function with a builder pattern
@@ -25,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Implement simple logger that prints messages to the output console in Godot
 
-[unreleased]: https://github.com/jdno/godot-logger/compare/v0.2.0...HEAD
+[unreleased]: https://github.com/jdno/godot-logger/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/jdno/godot-logger/releases/tag/v0.3.0
 [0.2.0]: https://github.com/jdno/godot-logger/releases/tag/v0.2.0
 [0.1.0]: https://github.com/jdno/godot-logger/releases/tag/v0.1.0